requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
markovify==0.9.4
pydub==0.25.1
python-telegram-bot==13.15
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
six==1.16.0
SpeechRecognition==3.9.0
tornado==6.1
tzdata==2022.7
tzlocal==4.2
Unidecode==1.3.6
urllib3==1.26.13
